// src/app/shared/model/host.ts
export class Host {
    hostid: string;
    host: string;
    available: number;
    description: string;
    disable_until: number;
    error: string;
    errors_from: number;
    flags: number;
    inventory_mode: number;
    ipmi_authtype: number;
    ipmi_available: number;
    ipmi_disable_until: number;
    ipmi_error: string;
    ipmi_errors_from: number;
    ipmi_password: string;
    ipmi_privilege: number;
    ipmi_username: string;
    jmx_available: number;
    jmx_disable_until: number;
    jmx_error: string;
    jmx_errors_from: number;
    maintenance_from: number;
    maintenance_status: number;
    maintenance_type: number;
    maintenanceid: string;
    name: string;
    proxy_hostid: string;
    snmp_available: number;
    snmp_disable_until: number;
    snmp_error: string;
    snmp_errors_from: number;
    status: number;
    tls_connect: number;
    tls_accept: number;
    tls_issuer: string;
    tls_subject: string;
    tls_psk_identity: string;
    tls_psk: string;
}