// src/app/shared/model/graph.ts
export class Graph {
    graphid: string ;
    height: number ;
    name: string ;
    width: number ;
    flags: number ;
    graphtype: number ;
    percent_left: boolean ;
    percent_right: boolean ;
    show_3d: number ;
    show_legend: number ;
    show_work_period: number ;
    templateid: string ;
    yaxismax: any ;
    yaxismin: any ;
    ymax_itemid: string ;
    ymax_type: number ;
    ymin_itemid: string;
    ymin_type: number ;

}